+++ R/bggAPI.R
@@ -67,7 +67,7 @@ bggAPI <- R6::R6Class(
     ),
 
     public = list(
-    # Fetch ----------------------------------------------------------------
+    # Fetch --------------------------------------------------------------------
     #' @description Fetches variables with given \code{variable_names} from the
     #'   object's \code{xml}. Returns them as a list. This is a main method of
     #'   getting non-scalar variables (as they are hard to fit into a
@@ -77,8 +77,8 @@ bggAPI <- R6::R6Class(
     #'   fetch.
     #' @param compress a logical value, decides whether the fetched variables
     #'   should be compress into a scalar form (if possible).
-    fetch = .fetch_external <- function(variable_names = NULL,
-                                        compress = FALSE)
+    fetch = function(variable_names = NULL,
+                     compress = FALSE)
     {
         # Assign to avoid NOTEs while checking the package
         Variable <- NULL
@@ -89,7 +89,10 @@ bggAPI <- R6::R6Class(
         specs <- private$.get_varspecs()
 
         if (!is.null(variable_names)) {
-            assert_that(.are_strings(variable_names))
+            assert_character(variable_names,
+                             any.missing = FALSE,
+                             min.len = 1,
+                             unique = TRUE)
         } else {
             variable_names <- specs$Variable
         }
@@ -111,10 +114,57 @@ bggAPI <- R6::R6Class(
             }
         }
 
-        result <- .fetch_internal(xml = private$.xml,
-                                  variable_names = variable_names,
-                                  var_specs = specs,
-                                  compress = compress)
+        # Loop for every variable ----------------------------------------------
+        xml <- private$.xml
+
+        result <- list()
+        for (var in variable_names) {
+            vs <- specs[Variable == var]
+
+            # Extract ----------------------------------------------------------
+            if (vs$Custom != "") {
+                fun <- .internal_fun(paste0(".fetch_", vs$Custom))
+                fetched <- fun(xml)
+            } else if (vs$Attribute != "") {
+                fun <- .internal_fun(paste0(".attr2", vs$Type))
+                fetched <- fun(xml = xml,
+                               xpath = vs$Node,
+                               attr = vs$Attribute,
+                               scalar = vs$Scalar)
+            } else {
+                fun <- .internal_fun(paste0(".nodes2", vs$Type))
+                fetched <- fun(xml = xml,
+                               xpath = vs$Node,
+                               scalar = vs$Scalar)
+            }
+
+            # Apply postprocessing function ------------------------------------
+            if (vs$Postprocessing != "") {
+                post_fun <- .internal_fun(paste0(".", vs$Postprocessing))
+
+                if (vs$Scalar) {
+                    fetched <- post_fun(fetched)
+                } else {
+                    fetched <- lapply(fetched, post_fun)
+                }
+            }
+
+            # Compression ------------------------------------------------------
+            if (compress) {
+                if (vs$Compression == "toString") {
+                    fetched <- sapply(fetched, toString)
+                }
+                if (vs$Compression == "squeeze") {
+                    fetched <- suppressWarnings(lapply(fetched, as.numeric))
+                    fetched <- sapply(fetched, squeeze)
+                }
+            }
+
+            result[[vs$Variable]] <- fetched
+        }
+
+        # Naming
+        names(result) <- variable_names
         return(result)
     },
 
@@ -136,7 +186,10 @@ bggAPI <- R6::R6Class(
         specs <- private$.get_varspecs()
 
         if (!is.null(variable_names)) {
-            assert_that(.are_strings(variable_names))
+            assert_character(variable_names,
+                             any.missing = FALSE,
+                             min.len = 1,
+                             unique = TRUE)
 
             # Already present and omitted
             existing <- intersect(variable_names, names(private$.data))
@@ -199,7 +252,7 @@ bggAPI <- R6::R6Class(
     switch_namestyle = function(to)
     {
         # Assertions
-        assert_that(.is_string(to, allowed = c("pretty", "classic")))
+        assert_choice(to, choices = c("pretty", "classic"))
 
         # Assign to avoid NOTEs while checking the package
         Class <- NULL