+++ R/custom_fetches.R
@@ -17,49 +17,58 @@ NULL
 #'
 .fetch_pollplayers <- function(xml)
 {
-    polls <- html_node(xml, xpath = "poll[@name = 'suggested_numplayers']")
-    polls <- lapply(polls, html_nodes, xpath = "results")
+    polls <- xml_find_first(xml, xpath = "poll[@name = 'suggested_numplayers']")
+    polls <- lapply(polls, xml_find_all, xpath = "results")
+    lng_polls <- .xml_concatenate(polls)
 
-    numplayers <- lapply(polls, .attr2text, xpath = ".",
-                         attr = "numplayers", scalar = TRUE)
-    best <- lapply(polls, .attr2number,
-                   xpath = "result[@value = 'Best']",
-                   attr = "numvotes", scalar = TRUE)
-    recommended <- lapply(polls, .attr2number,
-                          xpath = "result[@value = 'Recommended']",
-                          attr = "numvotes", scalar = TRUE)
-    not_recc <- lapply(polls, .attr2number,
-                       xpath = "result[@value = 'Not Recommended']",
-                       attr = "numvotes", scalar = TRUE)
+    numplayers <- .attr2text(lng_polls,
+                             xpath = ".",
+                             attr = "numplayers",
+                             scalar = TRUE)
+    best <- .attr2number(lng_polls,
+                         xpath = "result[@value = 'Best']",
+                         attr = "numvotes",
+                         scalar = TRUE)
+    recommended <- .attr2number(lng_polls,
+                                xpath = "result[@value = 'Recommended']",
+                                attr = "numvotes",
+                                scalar = TRUE)
+    not_recc <- .attr2number(lng_polls,
+                             xpath = "result[@value = 'Not Recommended']",
+                             attr = "numvotes",
+                             scalar = TRUE)
 
-    result <- list()
-    for (i in seq_along(polls)) {
-        result[[i]] <- data.table(numplayers = numplayers[[i]],
-                                  best = best[[i]],
-                                  recommended = recommended[[i]],
-                                  notrecommended = not_recc[[i]])
-    }
+    # Split result back by IDs
+    result <- data.table(numplayers = numplayers,
+                         best = best,
+                         recommended = recommended,
+                         notrecommended = not_recc)
+    result <- .split_acc2list(result, polls)
 
-    return (result)
+    return(result)
 }
 
 #' @describeIn custom_fetches Method for pollage of bggGames.
 #'
 .fetch_pollage <- function(xml)
 {
-    polls <- html_node(xml, xpath = "poll[@name = 'suggested_playerage']")
-    polls <- lapply(polls, html_nodes, xpath = "results/result")
+    polls <- xml_find_first(xml, xpath = "poll[@name = 'suggested_playerage']")
+    polls <- lapply(polls, xml_find_all, xpath = "results/result")
+    lng_polls <- .xml_concatenate(polls)
 
-    age <- lapply(polls, .attr2text, xpath = ".",
-                  attr = "value", scalar = TRUE)
-    votes <- lapply(polls, .attr2number, xpath = ".",
-                    attr = "numvotes", scalar = TRUE)
+    age <- .attr2text(lng_polls,
+                      xpath = ".",
+                      attr = "value",
+                      scalar = TRUE)
+    votes <- .attr2number(lng_polls,
+                          xpath = ".",
+                          attr = "numvotes",
+                          scalar = TRUE)
 
-    result <- list()
-    for (i in seq_along(polls)) {
-        result[[i]] <- data.table(age = age[[i]],
-                                  votes = votes[[i]])
-    }
+    # Split results by IDs
+    result <- data.table(age = age,
+                         votes = votes)
+    result <- .split_acc2list(result, polls)
 
     return(result)
 }
@@ -68,25 +77,41 @@ NULL
 #'
 .fetch_polllanguage <- function(xml)
 {
-    polls <- html_node(xml, xpath = "poll[@name = 'language_dependence']")
-    polls <- lapply(polls, html_nodes, xpath = "results/result")
+    polls <- xml_find_first(xml, xpath = "poll[@name = 'language_dependence']")
+    polls <- lapply(polls, xml_find_all, xpath = "results/result")
+    lng_polls <- .xml_concatenate(polls)
 
-    description <- lapply(polls, .attr2text, xpath = ".",
-                          attr = "value", scalar = TRUE)
-    level <- lapply(polls, .attr2number, xpath = ".",
-                    attr = "level", scalar = TRUE)
-    votes <- lapply(polls, .attr2number, xpath = ".",
-                    attr = "numvotes", scalar = TRUE)
+    description <- .attr2text(lng_polls,
+                              xpath = ".",
+                              attr = "value",
+                              scalar = TRUE)
+    level <- .attr2number(lng_polls,
+                          xpath = ".",
+                          attr = "level",
+                          scalar = TRUE)
+    votes <- .attr2number(lng_polls,
+                          xpath = ".",
+                          attr = "numvotes",
+                          scalar = TRUE)
 
-    result <- list()
-    for (i in seq_along(polls)) {
-        result[[i]] <- data.table(level = level[[i]],
-                                  description = description[[i]],
-                                  votes = votes[[i]])
-    }
+    # Split results by IDs
+    result <- data.table(level = level,
+                         description = description,
+                         votes = votes)
+    result <- .split_acc2list(result, polls)
 
     return(result)
 }
+
+#' @describeIn custom_fetches Method for ranks of bggGames.
+#'
+.fetch_ranks_gms <- function(xml)
+    .fetch_ranks(xml, "statistics/ratings/ranks")
+
+#' @describeIn custom_fetches Method for ranks of bggCollection.
+#'
+.fetch_ranks_cllctn <- function(xml)
+    .fetch_ranks(xml, "stats/rating/ranks")
 
 #' @describeIn custom_fetches Method for bestplayers of bggGames.
 #'
@@ -98,12 +123,14 @@ NULL
 #'
 .fetch_notrecplayers <- function(xml) .playerpoll_outcome(xml, "notrecommended")
 
+
+
 # Multipurpose #################################################################
 .playerpoll_outcome <- function(xml, category)
 {
     # Assertions
     types <- c("best", "recommended", "notrecommended")
-    assert_that(.is_string(category, allowed = types))
+    assert_choice(category, choices = types)
 
     res <- .fetch_pollplayers(xml)
 
@@ -120,4 +147,33 @@ NULL
     res <- split(res$outcome, factor(res$i, levels = seq_along(xml)))
     res <- unname(res)
     return(res)
+}
+
+.fetch_ranks <- function(xml, ranks_xpath)
+{
+    ranks <- xml_find_first(xml, xpath = ranks_xpath)
+    tabs <- lapply(ranks, function(x) xml_attrs(
+        xml_find_all(x, xpath = "./rank")))
+
+    # Assignment to avoid NOTEs while checking the package
+    friendlyname <- NULL
+    value <- NULL
+    bayesaverage <- NULL
+
+    # Join all ranks into one data.table
+    result <- rbindlist(lapply(unlist(tabs, recursive = FALSE), as.list))
+    result[, ":="(friendlyname = str_remove(friendlyname, " Rank$"),
+                  value = as.numeric(
+                      fifelse(value == "Not Ranked",
+                              NA_character_,
+                              value)),
+                  bayesaverage = as.numeric(
+                      fifelse(bayesaverage == "Not Ranked",
+                              NA_character_,
+                              bayesaverage)),
+                  id = NULL)]
+
+    # Split results per ID once again
+    result <- .split_acc2list(result, tabs)
+    return(result)
 }