+++ R/utils.R
@@ -16,9 +16,9 @@
 .compress <- function(vec, n_show = 5, collapse = ", ")
 {
     # Assertions
-    assert_that(is.atomic(vec))
-    assert_that(.is_positive_integer(n_show))
-    assert_that(.is_string(collapse))
+    assert_atomic(vec)
+    assert_count(n_show)
+    assert_string(collapse)
 
     n <- length(vec)
     extra <- n - n_show
@@ -29,7 +29,7 @@
         string <- paste0(string, "... (", extra, " more)")
     }
 
-    return (string)
+    return(string)
 }
 
 
@@ -46,13 +46,51 @@
 .plural <- function(string, count)
 {
     # Assertions
-    assert_that(.is_string(string))
-    assert_that(.is_integer(count))
+    assert_string(string)
+    assert_integerish(count, lower = 0)
 
     if (count > 1) {
         string <- paste0(string, "s")
     }
 
-    return (string)
+    return(string)
 }
 
+
+#' Split according to a list
+#'
+#' This splits \code{x} using \code{\link[base]{split}} by making sure that
+#' the returned list is of the same length as \code{list} and every element
+#' is of the same length as \code{list}'s elements.
+#'
+#' @param x an \code{R} object to split.
+#' @param list a list.
+#'
+#' @return A list.
+#' @keywords internal
+#'
+.split_acc2list <- function(x, list)
+{
+    assert_list(list)
+
+    vec <- seq_along(list)
+    lens <- lengths(list)
+    splitter <- factor(rep(vec, lens), levels = vec)
+    result <- unname(split(x, splitter))
+    return(result)
+}
+
+#' Get one of the internal functions from the package
+#'
+#' This makes sure internal functions from this package can be accessed which
+#' sometimes fails when using \code{match.fun}.
+#'
+#' @param fun_name a single string.
+#'
+#' @return A function.
+#' @keywords internal
+#'
+.internal_fun <- function(fun_name)
+{
+    get(fun_name, envir = asNamespace("bggAnalytics"))
+}