+++ R/zzz.R
@@ -1,10 +1,9 @@
 #' @keywords internal
 #'
-#' @import assertthat
+#' @import checkmate
 #' @import data.table
 #' @importFrom pryr unenclose
 #' @import R6
-#' @import rvest
 #' @import stringr
 #' @importFrom utils globalVariables
 #' @import xml2