+++ NEWS.md
@@ -1,6 +1,11 @@
-# bggAnalytics 0.1.1
+# bggAnalytics 0.2.0 (2021-09-23)
 
-* Fixed a bug which caused fetch variables to have improper names.
+* Slight optimization of variable fetching.
+* Added `ranks` and `families` and  variables to `bggGames` and `bggCollection` classes.
+* Added a `lastmodified` variable to the `bggCollection` class.
+* Improved the consistency of the `unsqueeze()` function output.
+* Fixed bugged calls to internal package functions.
+* Fixed a bug which caused `fetch()` variables to have improper names.
 
 # bggAnalytics 0.1.0 (2020-10-07)